refactor(front): extract function
